fix(toolset): hoist reify into universal pipeline
`reify_json_structured_content` previously fired only inside
`UpstreamToolSet::call`. Deployment-proxy upstreams (galoy_staging_*) and
any other custom toolset class skipped it, so the persisted invocation's
`original_structured` stayed empty even when raw_text was JSON-shaped.

Result: the walker entered as `Value::String`, fell back to byte-level
elision, and the new JSON queries (`json_path`, `json_array_slice`) hit
"invocation has no structured_content" errors.

Move the helper to a dedicated `reify` module and call it once per raw
result entering the classify/persist path:
- `call_top_level_tool` non-bypass branch (most MCP entry points)
- compose's `run_searchable_call` (always)
- compose's `run_top_level_call` (non-bypass branch)

Bypass tools (tool_output_fetch, compose itself) skip — they manage their
own pipeline. Tests follow the helper to its new location.
